pyproject.toml
[project]
name = "weird-ai-experiment-ideator"
version = "0.1.0"
description = "Multi-agent system for generating weird AI experiment ideas"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.12.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "markdown>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
